requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
langchain
pandas
numpy
python-dotenv
scipy
asyncio
Flask
Flask-Cors
openai
requests
gunicorn
nest_asyncio
faiss-cpu
selenium
beautifulsoup4
google-cloud-storage
google-auth
webdriver-manager
google-cloud-secret-manager